--- JejuDorang/src/styles/README.md
@@ -1,7 +1,0 @@
-# 🎨️ Styles
-
-여기는 JejuDorang Team의 styles directory 입니다.
-
----
-
-### 🗒️ Description
